+++ MercatorTechnologyTask/Mercator.Data.Infrastructure/Infrastruture/IGenericRepository.cs
@@ -9,7 +9,7 @@ namespace Mercator.Data.Domain.Infrastructure
     public interface IGenericRepository<T> where T : class
     {
         T Get(long id);
-        void Delete(int id);
+        void Delete(long id);
         void Add(T entity);
         List<T> GetAll();
         void Update(T entity);
